added leetcode 977

+++ interview_questions/two_pointers_squares_of_sorted_array.cpp
@@ -118,9 +118,9 @@ int Solution::subarraysWithKDistinct_bruteforce(vector<int>& A, int K)
         for(int j=i; j<len; j++)
         {
             s.insert(A[j]);
-            if(s.size() > K)
+            if((int)s.size() > K)
                 break;
-            else if(s.size() == K)
+            else if((int)s.size() == K)
                 ans++;
         }
     }
